# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: Seal HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /assistant/chat:
    post:
      tags: [assistant]
      summary: 流式对话（NDJSON）
      description: |
        请求头 X-Beeftv-Ui-Session 由 /assistant/ui-session 签发。响应为 NDJSON：
        若干 text_delta 增量，最后恰好一条 turn_end 结算（含 reply、toolCalls、error、cancelled）。
        宿主凭据与模型密钥由后端注入，浏览器不接触。
        X-Beeftv-Turn-Admission 响应头为 rejected（尚未尝试创建业务轮次）、unknown（已尝试创建）或 admitted（已创建）。
        只有 rejected 可以确认未接纳；缺少标记、网络故障或接纳后失败仍需读取回执。续签凭据不会重放消息。
      responses:
        "200":
          description: application/x-ndjson 事件流
        "403":
          description: 缺少或过期的 UI 会话凭据
        "503":
          description: 宿主未配置或不可达（reason=host_token_missing / host_unreachable）
  /assistant/cancel:
    post:
      tags: [assistant]
      summary: 取消指定画布正在进行的回合
      description: 需要写权限，并且 canvasId 必须属于当前工作区。只读客户端不能取消。
      responses:
        "202":
          description: 已受理取消
        "403":
          description: 只读客户端，或缺少写权限
        "404":
          description: 画布不存在或不属于当前工作区，或该画布没有正在进行的会话
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
...